+++ Sources/MapboxNavigation/CarPlayMapViewControllerDelegate.swift
@@ -11,8 +11,8 @@ public protocol CarPlayMapViewControllerDelegate: AnyObject, UnimplementedLoggin
      Tells the receiver that the final destination `PointAnnotation` was added to the `CarPlayMapViewController`.
      
      - parameter carPlayMapViewController: The `CarPlayMapViewController` object.
-     - parameter finalDestinationAnnotation: `PointAnnotation`, which was added to the `MapView`.
-     - parameter pointAnnotationManager: `PointAnnotationManager` instance, which is responsible for `PointAnnotation`s management in the `NavigationMapView`.
+     - parameter finalDestinationAnnotation: The point annotation that was added to the map view.
+     - parameter pointAnnotationManager: The object that manages the point annotation in the map view.
      */
     func carPlayMapViewController(_ carPlayMapViewController: CarPlayMapViewController,
                                   didAdd finalDestinationAnnotation: PointAnnotation,