+++ Sources/MapboxNavigation/CarPlayNavigationViewControllerDelegate.swift
@@ -29,8 +29,8 @@ public protocol CarPlayNavigationViewControllerDelegate: AnyObject, Unimplemente
      Tells the receiver that the final destination `PointAnnotation` was added to the `CarPlayNavigationViewController`.
      
      - parameter carPlayNavigationViewController: The `CarPlayNavigationViewController` object.
-     - parameter finalDestinationAnnotation: `PointAnnotation`, which was added to the `MapView`.
-     - parameter pointAnnotationManager: `PointAnnotationManager` instance, which is responsible for `PointAnnotation`s management in the `NavigationMapView`.
+     - parameter finalDestinationAnnotation: The point annotation that was added to the map view.
+     - parameter pointAnnotationManager: The object that manages the point annotation in the map view.
      */
     func carPlayNavigationViewController(_ carPlayNavigationViewController: CarPlayNavigationViewController,
                                          didAdd finalDestinationAnnotation: PointAnnotation,