+++ Sources/MapboxNavigation/NavigationMapViewDelegate.swift
@@ -104,8 +104,8 @@ public protocol NavigationMapViewDelegate: AnyObject, UnimplementedLogging {
      Tells the receiver that the final destination `PointAnnotation` was added to the `NavigationMapView`.
      
      - parameter navigationMapView: The `NavigationMapView` object.
-     - parameter finalDestinationAnnotation: `PointAnnotation`, which was added to the `MapView`.
-     - parameter pointAnnotationManager: `PointAnnotationManager` instance, which is responsible for `PointAnnotation`s management in the `NavigationMapView`.
+     - parameter finalDestinationAnnotation: The point annotation that was added to the map view.
+     - parameter pointAnnotationManager: The object that manages the point annotation in the map view.
      */
     func navigationMapView(_ navigationMapView: NavigationMapView, didAdd finalDestinationAnnotation: PointAnnotation, pointAnnotationManager: PointAnnotationManager)
 }